+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Ref: Return NoOpTransaction instead of null (#1126)
 * Fix: Do not crash when passing null values to @Nullable methods, eg User and Scope
 * Fix: Resolving dashed properties from external configuration
+* Feat: Read `uncaught.handler.enabled` property from the external configuration 
 
 # 4.0.0-alpha.2